Rename trait to `PackageSources`
